// src/app/models/negocio.ts
export class Negocio {
    key: string;
    nombre: string;
    categoria: string;
    administrador: string;
    subcategoriah: string;
    subcategoriar: string;
    subcategoriae: string;
    direccion: string;
    latitude: number;
    longitude:number;
    type:string;
    mascotas:string;
    posicion:string;
    portada:boolean;
    pais:string;
    departamento:string;
    ciudad:string;
}